Unified head interception
Motivation:

The new interceptors for changing request/response head each do a separate change yet they only modify the request/response head.

We should instead have a single interceptor that can implement all of them a batch the modification of the head in a single interceptor.

The builder pattern can be used to build such interceptor with the desired modification of the developer

Changes:

A new HeadInterceptor with a builder that aggregates the existing head individual interceptors.