Body transformer should accept and produce a media type
Motivation:

Currently body transformer are applied on the media type whatsoever.

A body transformer should declare the list of media type it produces as well as indicate which media type it consumes.

These should be taken in account when applying the body transformer or not on the proxied stream.

Changes:

Add a body transformer produces/consumes method.

The generic proxy interceptor uses the produces/consumes to decide whether or not to apply the transformer based on the response accept header.

The Body interface has been added a media type that indicates the media type of the body entity which can be used to match against transformers and also let the transformer set a body with a different media type.