Make the BodyTransformer asynchronous instead of being synchronous.
Motivation:

The design of BodyTransformer is synchronous as currently it only applies to buffering functions.

We should make this contract asynchronous and move the synchronous part in the implementations we currently support.

Changes:

BodyTransform transforms Body->Future<Body> instead of Buffer->Buffer

Synchronous transformations are moved to a BodyTransfomers class

Buffering overflow moves to the synchronous transformer instead of being in the interceptor since now the interceptor only sees an asynchronous transformation.