Introduce OriginRequestProvider (#123)
Closes #35, closes #44

OriginRequestProvider is a Vert.x API type that unifies all the origin selection methods currently present as overloaded methods in HttpProxy.

OriginRequestProvider being a FunctionalInterface, users can provide one in the form of a lambda.

OriginRequestProvider is a little different from the existing contract.
Instead of the proxied request and HTTP client as parameters, it has the ProxyContext (that exposes all the proxy data and the client).

This allows implementations to make a decision on the modifications interceptors can make (including when the request is an upgrade to WebSocket). See #44

It also allows interceptor and provider to exchange data via the ProxyContext attachments. See #35

Signed-off-by: Thomas Segismont <tsegismont@gmail.com>